+++ packages/sdk-core/src/subgraph/entities/poolMember/poolMember.ts
@@ -30,7 +30,6 @@ export interface PoolMember {
     totalAmountClaimed: BigNumber;
     token: Address;
     totalAmountReceivedUntilUpdatedAt: BigNumber;
-    poolTotalAmountDistributedUntilUpdatedAt: BigNumber;
     pool: Address;
     syncedPerUnitFlowRate: BigNumber;
     syncedPerUnitSettledValue: BigNumber;