+++ packages/sdk-core/CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 ### Changed
+- Remove `poolTotalAmountDistributedUntilUpdatedAt` from subgraph query
 ### Fixed
 
 ## [0.8.0] - 2024-08-01